requirements.txt
asgiref==3.4.1
attrs==21.2.0
coverage==5.5
Django==3.2.7
django-environ==0.7.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
Faker==8.12.1
iniconfig==1.1.1
mixer==7.1.2
packaging==21.0
pluggy==1.0.0
psycopg2-binary==2.9.1
py==1.10.0
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.5
pytest-cov==2.12.1
pytest-django==4.4.0
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
